augur: rename independent_exogenous->independent, exogenous_model->model
Both modules live where the redundancy is obvious: model/ is entirely about
modeling exogenous series, and augur/fit/model.py defines the model protocols
the fitter targets. The 'exogenous' qualifier was redundant.

Scoped to the module names (files + BUILD targets + import paths). Class names
(IndependentExogenousModel/Config), runtime label strings, and the serialized
exogenous_model_id wire field are deliberately left for a follow-up so this PR
is a pure, reviewable rename with no wire/provenance impact.
