+++ augur/fit/metrics_report.py
@@ -29,10 +29,10 @@ from typing import Any, cast
 
 from augur.api.config import load_augur_config
 from augur.fit.data import load_historical
-from augur.fit.exogenous_model import FittableScorable, Scorable
 from augur.fit.metrics import held_out_predictive_score, multi_step_predictive_score, rolling_origin_predictive_score
+from augur.fit.model import FittableScorable, Scorable
 from augur.model.exogenous_provider_config import CompositeExogenousProviderConfig
-from augur.model.independent_exogenous import IndependentExogenousProviderConfig
+from augur.model.independent import IndependentExogenousProviderConfig
 from augur.model.path_models.scenarios import HistoricalSeries
 from augur.model.vecm import VecmConfig, VecmModel
 from util.bazel.runfiles import get_required_path
