+++ augur/fit/model.py
