+++ augur/model/independent.py
