+++ augur/plans/typed_series_config.md
@@ -121,7 +121,7 @@ config-with-prefix surfaces). **Deferred to Phase 3:** 4, 5, 6 — see note belo
 
 | #   | File / class                                                                    | New shape                                                                                                                                                                                            | status    |
 | --- | ------------------------------------------------------------------------------- | ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | --------- |
-| 1   | `model/independent_exogenous.py` `IndependentExogenousProviderConfig`           | inherit `LevelSeriesGroups[ScalarSeriesSpec]` (per-kind fields, no `series:` wrapper) + `type` + `private_equity_marks: dict[IssuerId, ScalarSeriesSpec]`; drop `_classified_series` prefix re-parse | ✅ done   |
+| 1   | `model/independent.py` `IndependentExogenousProviderConfig`                     | inherit `LevelSeriesGroups[ScalarSeriesSpec]` (per-kind fields, no `series:` wrapper) + `type` + `private_equity_marks: dict[IssuerId, ScalarSeriesSpec]`; drop `_classified_series` prefix re-parse | ✅ done   |
 | 2   | `model/series_model.py` `IndependentSeriesModels` (sim/bench twin)              | same collapse via `LevelSeriesGroups.from_level_keys`                                                                                                                                                | ✅ done   |
 | 3   | `api/portfolio.py` `HoldingPositionConfig`                                      | `value_series: AssetKey` typed in YAML; drop `value_series_id` + `asset_key` re-parse; `_validate_references`/`level_anchors`/`to_initial_lots` consume the typed key                                | ✅ done   |
 | 7   | `model/sample_sanity.py` checks                                                 | `key: LevelSeriesKey`, `required_level_series: tuple[LevelSeriesKey,...]`, `issuer_id: IssuerId`                                                                                                     | ✅ done   |
